Small changes here and there
