+++ frontend/markbyte/src/App.jsx
@@ -3,9 +3,7 @@ import {
   BrowserRouter as Router,
   Routes,
   Route,
-  useParams,
 } from "react-router-dom";
-import axios from "axios";
 import { useEffect, useState } from "react";
 import Home from "./components/pages/Home/Home";
 import Auth from "./components/pages/Auth/Auth";
@@ -15,36 +13,10 @@ import BloggerLandingPage from "./components/pages/BloggerLanding/BloggerLanding
 import BloggerHome from "./components/pages/BloggerHome/BloggerHome";
 import EditorPreview from "./components/pages/LivePreview/CreateEditorPreview";
 import PublishEditorPreview from "./components/pages/LivePreview/PublishEditorPreview";
+import DynamicBlogPost from "./components/pages/Post/DynamicBlogPost";
 import { useAuth } from "./contexts/AuthContext";
 import { Toaster } from "./components/ui/toaster";
 import "./App.css";
-
-// function Footer() {
-//   return (
-//     <footer className="footer">
-//       <p>&copy; Group 11. Rishab Pangal, Anish Laddha, Shrijan Swaminathan</p>
-//     </footer>
-//   );
-// }
-
-function DynamicBlogPost() {
-  const { user, post } = useParams();
-
-  useEffect(() => {
-    axios
-      .get(`http://localhost:8080/${user}/${post}`)
-      .then((response) => {
-        document.open();
-        document.write(response.data);
-        document.close();
-      })
-      .catch((error) => {
-        console.error("Error fetching blogger's post page:", error);
-      });
-  }, [user, post]);
-
-  return null;
-}
 
 function App() {
   const { isAuthenticated } = useAuth();
@@ -67,7 +39,7 @@ function App() {
             ) : !isAuthenticated ? (
               <Home />
             ) : (
-              <BloggerHome /> 
+              <BloggerHome />
             )
           }
         />
@@ -78,7 +50,7 @@ function App() {
             isLoading ? (
               console.log("Loading...")
             ) : !isAuthenticated ? (
-              <Home/>
+              <Home />
             ) : (
               <EditorPreview />
             )
@@ -90,9 +62,9 @@ function App() {
             isLoading ? (
               console.log("Loading...")
             ) : !isAuthenticated ? (
-              <Home/>
+              <Home />
             ) : (
-              < PublishEditorPreview/>
+              <PublishEditorPreview />
             )
           }
         />
