+++ frontend/markbyte/src/components/pages/BloggerHome/BloggerHome.jsx
@@ -469,7 +469,7 @@ function BloggerHome() {
               <CardContent className="px-7 pb-7 flex flex-col justify-end h-full">
                 <div>
                   <p className="text-gray-500 font-medium mb-2">
-                    Upload your file
+                    Upload your blog files
                   </p>
                   <button
                     className="mt-3 bg-gradient-to-r from-[#005a7a] to-[#084464] text-white font-medium py-3.5 px-6 rounded-xl transition-all duration-300 w-full flex items-center justify-center gap-2 hover:shadow-md hover:translate-y-[-2px] active:translate-y-[0px]"
